requirements.txt
# Core ML / AI
torch==2.11.0
transformers==5.5.0
safetensors==0.7.0
tokenizers==0.22.2
huggingface_hub==1.9.1
hf_xet==1.4.3

# Numerical / Scientific
numpy==2.4.4
sympy==1.14.0
mpmath==1.3.0
networkx==3.6.1
regex==2026.4.4

# Database
psycopg2-binary==2.9.11

# Geo
geopy==2.4.1
geographiclib==2.1

# HTTP / Async
httpx==0.28.1
httpcore==1.0.9
anyio==4.13.0
h11==0.16.0
certifi==2026.2.25
idna==3.11

# Scheduling
apscheduler==3.11.2
tzlocal==5.3.1
tzdata==2026.1

# CLI / Terminal
typer==0.24.1
click==8.3.2
rich==14.3.3
pygments==2.20.0
colorama==0.4.6
shellingham==1.5.4

# Templating / Serialization
jinja2==3.1.6
markupsafe==3.0.3
pyyaml==6.0.3

# Utilities
python-dotenv==1.2.2
tqdm==4.67.1
filelock==3.25.2
fsspec==2026.3.0
packaging==26.0
typing_extensions==4.15.0
feedparser==6.0.12
sgmllib3k==1.0.0
markdown-it-py==4.0.0
mdurl==0.1.2
annotated_doc==0.0.4

# Dashboard
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.0.0
